api/codebases: move in-memory repo into db package
Created with Sturdy
