+++ api/pkg/codebases/db/codebase_user.go
@@ -2,6 +2,7 @@ package db
 
 import (
 	"context"
+	"database/sql"
 	"fmt"
 
 	"getsturdy.com/api/pkg/codebases"
@@ -76,6 +77,71 @@ func (r *codebaseUserRepo) DeleteByID(ctx context.Context, id string) error {
 	_, err := r.db.ExecContext(ctx, `DELETE FROM codebase_users WHERE id = $1`, id)
 	if err != nil {
 		return fmt.Errorf("failed to delete codebase_users by id: %w", err)
+	}
+	return nil
+}
+
+type inMemoryCodebaseUserRepository struct {
+	users []codebases.CodebaseUser
+}
+
+func NewInMemoryCodebaseUserRepo() CodebaseUserRepository {
+	return &inMemoryCodebaseUserRepository{}
+}
+
+func (r *inMemoryCodebaseUserRepository) GetByID(_ context.Context, id string) (*codebases.CodebaseUser, error) {
+	for _, u := range r.users {
+		if u.ID == id {
+			return &u, nil
+		}
+	}
+	return nil, sql.ErrNoRows
+}
+
+func (r *inMemoryCodebaseUserRepository) Create(entity codebases.CodebaseUser) error {
+	r.users = append(r.users, entity)
+	return nil
+}
+
+func (r *inMemoryCodebaseUserRepository) GetByUser(userID users.ID) ([]*codebases.CodebaseUser, error) {
+	var res []*codebases.CodebaseUser
+	for _, u := range r.users {
+		if u.UserID == userID {
+			u2 := u
+			res = append(res, &u2)
+		}
+	}
+	return res, nil
+}
+
+func (r *inMemoryCodebaseUserRepository) GetByCodebase(codebaseID codebases.ID) ([]*codebases.CodebaseUser, error) {
+	var res []*codebases.CodebaseUser
+	for _, u := range r.users {
+		if u.CodebaseID == codebaseID {
+			u2 := u
+			res = append(res, &u2)
+		}
+	}
+	return res, nil
+}
+
+func (r *inMemoryCodebaseUserRepository) GetByUserAndCodebase(userID users.ID, codebaseID codebases.ID) (*codebases.CodebaseUser, error) {
+	for _, u := range r.users {
+		if u.UserID == userID && u.CodebaseID == codebaseID {
+			return &u, nil
+		}
+	}
+	return nil, sql.ErrNoRows
+}
+
+func (r *inMemoryCodebaseUserRepository) DeleteByID(_ context.Context, id string) error {
+	for i, u := range r.users {
+		if u.ID == id {
+			// Remove the element at index i from a.
+			r.users[i] = r.users[len(r.users)-1]               // Copy last element to index i.
+			r.users[len(r.users)-1] = codebases.CodebaseUser{} // Erase last element (write zero value).
+			r.users = r.users[:len(r.users)-1]                 // Truncate slice.
+		}
 	}
 	return nil
 }
