+++ api/pkg/codebases/graphql/codebase_test.go
@@ -7,9 +7,9 @@ import (
 	"getsturdy.com/api/pkg/auth"
 	service_auth "getsturdy.com/api/pkg/auth/service"
 	"getsturdy.com/api/pkg/codebases"
+	db_codebase "getsturdy.com/api/pkg/codebases/db"
 	service_codebase "getsturdy.com/api/pkg/codebases/service"
 	"getsturdy.com/api/pkg/graphql/resolvers"
-	"getsturdy.com/api/pkg/internal/inmemory"
 	"getsturdy.com/api/pkg/users"
 
 	"github.com/google/uuid"
@@ -19,8 +19,8 @@ import (
 )
 
 func TestCodebaseAccess(t *testing.T) {
-	codebaseRepo := inmemory.NewInMemoryCodebaseRepo()
-	codebaseUserRepo := inmemory.NewInMemoryCodebaseUserRepo()
+	codebaseRepo := db_codebase.NewMemory()
+	codebaseUserRepo := db_codebase.NewInMemoryCodebaseUserRepo()
 	codebaseService := service_codebase.New(codebaseRepo, codebaseUserRepo, nil, nil, nil, nil, nil, nil, nil, nil)
 	authService := service_auth.New(codebaseService, nil, nil, nil, nil, nil)
 	resolver := NewCodebaseRootResolver(
